scripts and trigger set up
