+++ Assets/Scripts/Systems/CheckPoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2417739fdf57724bb308777243a6586
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
