+++ Assets/Scripts/Systems/GameObjectSaveData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af9dcca2a4229ee468861b9cfb2a23b0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
